+++ packages/cactus-test-plugin-htlc-eth-besu/src/test/typescript/integration/plugin-htlc-eth-besu/get-status-endpoint.test.ts
@@ -221,12 +221,12 @@ test(testCase, async (t: Test) => {
   );
 
   const ids = [responseTxId.callOutput as string];
-  const res = await api.getStatusV1(
+  const res = await api.getStatusV1({
     ids,
     web3SigningCredential,
     connectorId,
     keychainId,
-  );
+  });
   t.equal(res.status, 200, "response status is 200 OK");
   t.equal(res.data[0], "1", "the contract status is 1 - Active");
   t.end();